bump(net-dns/libidn): sync with upstream

+++ net-dns/libidn/libidn-1.30.ebuild
@@ -1,18 +1,17 @@
-# Copyright 1999-2014 Gentoo Foundation
+# Copyright 1999-2015 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/net-dns/libidn/libidn-1.28.ebuild,v 1.12 2014/04/06 15:11:33 vapier Exp $
+# $Id$
 
 EAPI=5
-
-inherit elisp-common java-pkg-opt-2 mono-env
+inherit elisp-common java-pkg-opt-2 mono-env multilib-minimal
 
 DESCRIPTION="Internationalized Domain Names (IDN) implementation"
-HOMEPAGE="http://www.gnu.org/software/libidn/"
+HOMEPAGE="https://www.gnu.org/software/libidn/"
 SRC_URI="mirror://gnu/libidn/${P}.tar.gz"
 
 LICENSE="GPL-2 GPL-3 LGPL-3 java? ( Apache-2.0 )"
 SLOT="0"
-KEYWORDS="alpha amd64 arm arm64 hppa ia64 m68k ~mips ppc ppc64 s390 sh sparc x86 ~amd64-fbsd ~sparc-fbsd ~x86-fbsd ~x86-freebsd ~x86-interix ~amd64-linux ~ia64-linux ~x86-linux ~ppc-macos ~x64-macos ~x86-macos ~sparc-solaris ~sparc64-solaris ~x64-solaris ~x86-solaris"
+KEYWORDS="alpha amd64 arm ~arm64 hppa ia64 ~m68k ~mips ppc ppc64 ~s390 ~sh sparc x86 ~amd64-fbsd ~sparc-fbsd ~x86-fbsd ~x86-freebsd ~x86-interix ~amd64-linux ~ia64-linux ~x86-linux ~ppc-macos ~x64-macos ~x86-macos ~m68k-mint ~sparc-solaris ~sparc64-solaris ~x64-solaris ~x86-solaris"
 IUSE="doc emacs java mono nls static-libs"
 
 DOCS=( AUTHORS ChangeLog FAQ NEWS README THANKS TODO )
@@ -28,8 +27,12 @@ DEPEND="${COMMON_DEPEND}
 	)
 "
 RDEPEND="${COMMON_DEPEND}
-	nls? ( virtual/libintl )
+	nls? ( >=virtual/libintl-0-r1[${MULTILIB_USEDEP}] )
 	java? ( >=virtual/jre-1.5 )
+	abi_x86_32? (
+		!<=app-emulation/emul-linux-x86-baselibs-20140508-r5
+		!app-emulation/emul-linux-x86-baselibs[-abi_x86_32(-)]
+	)
 "
 
 pkg_setup() {
@@ -42,10 +45,11 @@ src_prepare() {
 	rm "${S}/java/${P}.jar" || die
 }
 
-src_configure() {
+multilib_src_configure() {
+	ECONF_SOURCE=${S} \
 	econf \
-		$(use_enable java) \
-		$(use_enable mono csharp mono) \
+		$(multilib_native_use_enable java) \
+		$(multilib_native_use_enable mono csharp mono) \
 		$(use_enable nls) \
 		$(use_enable static-libs static) \
 		--disable-silent-rules \
@@ -56,36 +60,44 @@ src_configure() {
 		--with-packager="Gentoo"
 }
 
-src_compile() {
+multilib_src_compile() {
 	default
 
-	if use emacs; then
-		elisp-compile src/*.el || die
+	if multilib_is_native_abi && use emacs; then
+		elisp-compile "${S}"/src/*.el || die
 	fi
 }
 
-src_install() {
-	default
+multilib_src_test() {
+	# only run libidn specific tests and not gnulib tests (bug #539356)
+	emake -C tests check
+}
 
-	if use emacs; then
-		# *.el are installed by the build system
-		elisp-install ${PN} src/*.elc
-		elisp-site-file-install "${FILESDIR}/50${PN}-gentoo.el"
-	else
-		rm -r "${ED}/usr/share/emacs" || die
-	fi
+multilib_src_install() {
+	emake DESTDIR="${D}" install
 
-	if use doc ; then
-		dohtml -r doc/reference/html/*
-	fi
-
-	if use java ; then
+	if multilib_is_native_abi && use java; then
 		java-pkg_newjar java/${P}.jar ${PN}.jar || die
 		rm -r "${ED}"/usr/share/java || die
 
 		if use doc ; then
 			java-pkg_dojavadoc doc/java
 		fi
+	fi
+}
+
+multilib_src_install_all() {
+	if use emacs; then
+		# *.el are installed by the build system
+		elisp-install ${PN} "${S}"/src/*.elc
+		elisp-site-file-install "${FILESDIR}/50${PN}-gentoo.el"
+	else
+		rm -r "${ED}/usr/share/emacs" || die
+	fi
+
+	einstalldocs
+	if use doc ; then
+		dohtml -r doc/reference/html/.
 	fi
 
 	prune_libtool_files
